+++ src/compile_tests.rs
@@ -28,11 +28,11 @@ mod borrows {
 
     /// ```compile_fail,E0502
     /// use vectree::VecTree;
-    ///                                          // |
-    /// let mut tree = VecTree::<String>::new(); // | let a = tree.get(1);
-    /// let a = tree.get(1);                     // |         ---- immutable borrow occurs here
-    /// for mut inode in tree.iter_depth_mut() { // | for mut inode in tree.iter_depth_mut() { // cannot borrow `tree` as mutable because it is also borrowed as immutable
-    ///     *inode = inode.to_uppercase();       // |                  ^^^^^^^^^^^^^^^^^^^^^ mutable borrow occurs here
+    ///                                               // |
+    /// let mut tree = VecTree::<String>::new();      // | let a = tree.get(1);
+    /// let a = tree.get(1);                          // |         ---- immutable borrow occurs here
+    /// for mut inode in tree.iter_post_depth_mut() { // | for mut inode in tree.iter_post_depth_mut() { // cannot borrow `tree` as mutable because it is also borrowed as immutable
+    ///     *inode = inode.to_uppercase();            // |                  ^^^^^^^^^^^^^^^^^^^^^^^^ mutable borrow occurs here
     /// }
     /// assert_eq!(a, "A");
     /// ```
@@ -42,12 +42,12 @@ mod borrows {
     /// use std::ops::Deref;
     /// use vectree::VecTree;
     ///
-    /// let mut tree = VecTree::<i32>::new();    // | let mut iter = tree.iter_depth_simple();
-    /// let mut iter = tree.iter_depth_simple(); // |                ---- immutable borrow occurs here
-    /// let a1_borrowed = iter.next().unwrap();  // | let a1_borrowed = iter.next().unwrap();
-    /// tree.clear();                            // | tree.clear();
-    /// let value = a1_borrowed.deref();         // | ^^^^^^^^^^^^ mutable borrow occurs here
-    /// println!("value: {value}");              // | let value = a1_borrowed.deref();
+    /// let mut tree = VecTree::<i32>::new();         // | let mut iter = tree.iter_post_depth_simple();
+    /// let mut iter = tree.iter_post_depth_simple(); // |                ---- immutable borrow occurs here
+    /// let a1_borrowed = iter.next().unwrap();       // | let a1_borrowed = iter.next().unwrap();
+    /// tree.clear();                                 // | tree.clear();
+    /// let value = a1_borrowed.deref();              // | ^^^^^^^^^^^^ mutable borrow occurs here
+    /// println!("value: {value}");                   // | let value = a1_borrowed.deref();
     /// ```
     fn must_not_compile4() {}
 
@@ -55,8 +55,8 @@ mod borrows {
     /// use std::ops::Deref;
     /// use vectree::VecTree;
     ///
-    /// let mut tree = VecTree::<i32>::new();   // | let mut iter = tree.iter_depth();
-    /// let mut iter = tree.iter_depth();       // |                ---- immutable borrow occurs here
+    /// let mut tree = VecTree::<i32>::new();   // | let mut iter = tree.iter_post_depth();
+    /// let mut iter = tree.iter_post_depth();  // |                ---- immutable borrow occurs here
     /// let a1_borrowed = iter.next().unwrap(); // | let a1_borrowed = iter.next().unwrap();
     /// tree.clear();                           // | tree.clear();
     /// let value = a1_borrowed.deref();        // | ^^^^^^^^^^^^ mutable borrow occurs here
@@ -69,8 +69,8 @@ mod borrows {
     /// use vectree::VecTree;
     ///
     /// let mut tree = VecTree::<i32>::new();
-    /// let mut iter = tree.iter_depth();       // |
-    /// let _a1 = iter.next();                  // | let mut iter = tree.iter_depth();
+    /// let mut iter = tree.iter_post_depth();  // |
+    /// let _a1 = iter.next();                  // | let mut iter = tree.iter_post_depth();
     /// let _a2 = iter.next();                  // |                ---- immutable borrow occurs here
     /// let a = iter.next().unwrap();           //
     /// let a1_borrowed = a.iter_children()     // | tree.clear();                // OK: doesn't compile
@@ -86,8 +86,8 @@ mod borrows {
     /// use vectree::VecTree;
     ///
     /// let mut tree = VecTree::<String>::new();
-    /// let mut iter = tree.iter_depth();       // |
-    /// let _a1 = iter.next();                  // | let mut iter = tree.iter_depth();
+    /// let mut iter = tree.iter_post_depth();  // |
+    /// let _a1 = iter.next();                  // | let mut iter = tree.iter_post_depth();
     /// let _a2 = iter.next();                  // |                ---- immutable borrow occurs here
     /// let a = iter.next().unwrap();           //
     /// let a1_borrowed = a.iter_children()     // | let a1_mut = tree.get_mut(4);
@@ -105,12 +105,12 @@ mod borrows {
     /// use vectree::VecTree;
     ///
     /// let mut tree = VecTree::<i32>::new();
-    /// let mut iter = tree.iter_depth_mut();    // |
-    /// let _a1 = iter.next().unwrap();          // | let mut iter = tree.iter_depth_mut();
-    /// let _a2 = iter.next();                   // |                ---- first mutable borrow occurs here
-    /// let a_mut = iter.next().unwrap();        //
-    /// let a1_borrowed = a_mut.iter_children()  // | tree.clear();
-    ///     .next().unwrap();                    // | ^^^^ second mutable borrow occurs here
+    /// let mut iter = tree.iter_post_depth_mut();    // |
+    /// let _a1 = iter.next().unwrap();               // | let mut iter = tree.iter_post_depth_mut();
+    /// let _a2 = iter.next();                        // |                ---- first mutable borrow occurs here
+    /// let a_mut = iter.next().unwrap();             //
+    /// let a1_borrowed = a_mut.iter_children()       // | tree.clear();
+    ///     .next().unwrap();                         // | ^^^^ second mutable borrow occurs here
     /// tree.clear();
     /// let value = a1_borrowed.deref();
     /// println!("value: {value}");
@@ -121,9 +121,9 @@ mod borrows {
     /// use vectree::VecTree;
     ///
     /// let mut tree = VecTree::<String>::new();
-    /// let mut iter = tree.iter_depth_mut();
+    /// let mut iter = tree.iter_post_depth_mut();
     /// let mut a1_mut = iter.next().unwrap();   // |
-    /// let a1_borrowed = tree.get(4);           // | let mut iter = tree.iter_depth_mut();
+    /// let a1_borrowed = tree.get(4);           // | let mut iter = tree.iter_post_depth_mut();
     /// let value1 = a1_borrowed.clone();        // |                ---- mutable borrow occurs here
     /// *a1_mut = "new a1".to_string();          //
     /// let value2 = a1_borrowed.clone();        // | let a1_borrowed = tree.get(4);